fix(scripts): release.ts unintentionally also ran scripts
The release script used to import the utility `getPackages` from another script
called `run-script.ts`.

This in turn led to the side effect of `runInPackages` which would run
a script in all packages, if the first argument of release.ts happens to
be a script in that package.

It also led to a confusing error message on release.ts with no params:

```
% ./scripts/release.ts
Error: Must specify script name
```

This error is thrown in `run-scripts.ts`.

This PR moves the utilities to a different package with no side effects
on import.
