+++ scripts/release.ts
@@ -8,7 +8,7 @@ import parseArgs = require("minimist")
 import deline = require("deline")
 import { resolve, relative } from "path"
 import { readFile, createWriteStream, writeFile } from "fs-extra"
-import { getPackages } from "./run-script"
+import { getPackages } from "./script-utils"
 import Bluebird = require("bluebird")
 const replace = require("replace-in-file")
